docs: Added cache invalidation recipe to cookbook
